+++ src/styles/components.css
@@ -1272,6 +1272,28 @@
   gap: 6px;
 }
 
+.album-track-toolbar {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+  padding: 0 16px 8px;
+  flex-wrap: wrap;
+}
+
+.album-track-toolbar-filter {
+  position: relative;
+  flex: 1 1 160px;
+  max-width: 260px;
+}
+
+.album-track-toolbar-actions {
+  margin-left: auto;
+  display: flex;
+  align-items: center;
+  gap: var(--space-2);
+  min-height: 34px;
+}
+
 .album-info-dot {
   margin: 0 4px;
 }
